% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{calculate_index}
\alias{calculate_index}
\title{General function for index calculation based on a known dataset (i.e. microglia development index)}
\usage{
calculate_index(
  mat,
  index_df,
  gene_column = "gene",
  direction_column = "direction"
)
}
\arguments{
\item{mat}{A gene expression matrix}

\item{index_df}{Dataframe with columns including gene names, valence, and directionality}

\item{gene_column}{Name of the column in the dataframe that contains the gene names for indexing}

\item{direction_column}{Dataframe with columns including gene names, valence, and directionality}
}
\value{
Dataframe with index score for each sample (row from initial matrix)
}
\description{
General function for index calculation based on a known dataset (i.e. microglia development index)
}
