% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{calculate_index_deseq}
\alias{calculate_index_deseq}
\title{Function wrapper for calculating index using general calculate_index function on deseq2 object}
\usage{
calculate_index_deseq(
  dds,
  index_df,
  gene_column = "gene",
  direction_column = "direction",
  use_vst = TRUE,
  column_name = "developmental_index"
)
}
\arguments{
\item{dds}{A deseq2 dataset object already created}

\item{index_df}{Dataframe with columns including gene names, valence, and directionality}

\item{gene_column}{Name of the column in the dataframe that contains the gene names for indexing}

\item{direction_column}{Dataframe with columns including gene names, valence, and directionality}

\item{use_vst}{Conditional to use variance stabilizing transformation data or not. If not, it will default to calculating index on the raw gene counts}

\item{column_name}{What to name the output metadata column that is contained within the seurat object}
}
\value{
A deseq2 dataset obj that was the input object + a metadata column containing index values accessible by "column_name" argument in metadata
}
\description{
Function wrapper for calculating index using general calculate_index function on deseq2 object
}
