% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{calculate_index_seurat}
\alias{calculate_index_seurat}
\title{Function wrapper for calculating index using general calculate_index function on seurat object}
\usage{
calculate_index_seurat(
  seurat_obj,
  index_df,
  gene_column = "gene",
  direction_column = "direction",
  column_name = "developmental_index"
)
}
\arguments{
\item{seurat_obj}{A seurat object already created}

\item{index_df}{Dataframe with columns including gene names, valence, and directionality}

\item{gene_column}{Name of the column in the dataframe that contains the gene names for indexing}

\item{direction_column}{Dataframe with columns including gene names, valence, and directionality}

\item{column_name}{What to name the output metadata column that is contained within the seurat object}
}
\value{
A seurat obj that was the input object + a metadata column containing index values accessible by "column_name" argument in metadata
}
\description{
Function wrapper for calculating index using general calculate_index function on seurat object
}
