% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{calculate_tpm.DESeqDataSet}
\alias{calculate_tpm.DESeqDataSet}
\title{Function for calculating TPM on any gene matrix}
\usage{
calculate_tpm.DESeqDataSet(
  object,
  length_vector,
  assay_name = "TPM",
  log = FALSE,
  add_to_object = TRUE
)
}
\arguments{
\item{object}{DESeq2 object}

\item{length_vector}{Labelled character vector containing gene names and their lengths}

\item{assay_name}{What to name the resulting assay that is saved in the deseq2 object}

\item{log}{Whether to log transform the output or not (I usually don't)}

\item{add_to_object}{Conditional for adding to object. You should almost always do it}
}
\value{
A matrix of TPM values
}
\description{
Function for calculating TPM on any gene matrix
}
