% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{calculate_tpm.matrix}
\alias{calculate_tpm.matrix}
\title{Function for calculating TPM on any gene matrix}
\usage{
calculate_tpm.matrix(counts_mat, length_vector, log = FALSE)
}
\arguments{
\item{counts_mat}{Matrix of gene counts}

\item{length_vector}{Labelled character vector containing gene names and their lengths}

\item{log}{Whether to log transform the output or not (I usually don't)}
}
\value{
A matrix of TPM values
}
\description{
Function for calculating TPM on any gene matrix
}
