% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{find_transition_genes}
\alias{find_transition_genes}
\title{Function to find genes that are involved in transitions between cells across index values}
\usage{
find_transition_genes(
  s_obj,
  index_name = "developmental_index",
  chunk_size = 500,
  min_pct = 0.05
)
}
\arguments{
\item{s_obj}{A seurat object already created}

\item{index_name}{Which index to use for the calculations, contained in seurat object}

\item{chunk_size}{For chunking by x number of genes to do calculation without overloading memory, default of 500 genes works just fine}

\item{min_pct}{Minimum percent of cells that express a given gene. Default is 5\%, if a gene is expressed in less that 5\% of cells, it is dropped}
}
\value{
A dataframe containing the correlation between each gene and overall developmental index
}
\description{
Function to find genes that are involved in transitions between cells across index values
}
