% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsi_index}
\alias{fsi_index}
\title{Fast-Spiking Interneuron Index}
\format{
A data frame with columns:
\describe{
\item{Gene}{Gene symbol}
\item{direction}{Direction of regulation: "UP" or "DOWN"}
\item{valence}{Magnitude of change}
}
}
\usage{
fsi_index
}
\description{
A gene signature for scoring fast-spiking interneuron transcriptional identity.
}
\keyword{datasets}
