% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lps_index}
\alias{lps_index}
\title{LPS Inflammatory Microglia Index}
\format{
A data frame with columns:
\describe{
\item{gene}{Gene symbol}
\item{direction}{Direction of regulation: "UP" or "DOWN"}
\item{valence}{Magnitude of change}
}
}
\usage{
lps_index
}
\description{
A gene signature for scoring LPS-induced microglial inflammatory activation.
}
\keyword{datasets}
