% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdi}
\alias{mdi}
\title{Microglia Development Index}
\format{
A data frame with columns:
\describe{
\item{gene}{Gene symbol}
\item{direction}{Direction of regulation: "UP" or "DOWN"}
\item{valence}{Magnitude of change}
}
}
\usage{
mdi
}
\description{
A gene signature for scoring microglial developmental maturation state.
}
\keyword{datasets}
