% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{rolling_variance_index}
\alias{rolling_variance_index}
\title{Function for calculating rolling variance averaged across a specified gene window size}
\usage{
rolling_variance_index(
  s_obj,
  index_col = "developmental_index",
  variance_col = "gene_variance",
  window_size = 200
)
}
\arguments{
\item{s_obj}{A seurat object already created}

\item{index_col}{Name of the metadata column in the seurat object that contains index calculation}

\item{variance_col}{Name of the metadata column in the seurat object that contains gene variance calculation}

\item{window_size}{Window size (number of genes) to bin}
}
\value{
A dataframe containing the rolling index and rolling variance for comparison
}
\description{
Function for calculating rolling variance averaged across a specified gene window size
}
