% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculations.R
\name{seurat_gene_var}
\alias{seurat_gene_var}
\title{Function for calculating gene variance and adding it as a metadata column to the seurat object}
\usage{
seurat_gene_var(
  s_obj,
  assay = "RNA",
  slot = "data",
  column_name = "gene_variance"
)
}
\arguments{
\item{s_obj}{A seurat object already created}

\item{assay}{What to name the output metadata column that is contained within the seurat object}

\item{slot}{to name the output metadata column that is contained within the seurat object}

\item{column_name}{What to name the output metadata column that is contained within the seurat object}
}
\value{
A seurat obj that was the input object + a metadata column containing index values accessible by "column_name" argument in metadata
}
\description{
Function for calculating gene variance and adding it as a metadata column to the seurat object
}
